[NEW] Auto-TLS upgrade - tls connections have traditionally depended on the `tls` option being set. This option is multi-valued (undefined|boolean|tlsOptions) is problematic because it's meaning is complex:
If not set or false, connecting to a tls server fails
If set to true or tlsOptions, connecting to non-tls server fails

In the case where the certificates are not self-signed, the CA will need to be provided, but in the case the connection is expecting to be TLS from a specific CA.

If the value is not set, the connection should be allowed to be bare or tls.

It would be better if the protocol for the connection specified what to expect from the server but unfortunately this also has issues, as gossiped servers get a protocol specified by the client and not by the server doing the gossiping.

This change relaxes the situation where the `tls` connection option is not set and thus allows the client to connect to a bare or tls secured server. While still rejecting or requiring tls if the property is set.